feat: #678 - added bottom navigation bar to product page
The previous bottom navigation bottom bar was refactored, in order to be reused anywhere.
Then it was added to the product page.

Deleted files:
* `home_page.dart`
* `smooth_ui_library/smooth_bottom_navigation_bar.dart`
* `smooth_ui_library/smooth_bottom_navigation_bar_item.dart`

New file:
* `smooth_bottom_navigation_bar.dart`

Impacted files:
* `continuous_scan_page.dart`: added an explicit bottom navigation bar (was implicit through `HomePage`)
* `main.dart`: not using `HomePage` anymore; asking the bottom navigation bar who's the current page instead
* `new_product_page.dart`: added an explicit bottom navigation bar (was not there)
* `product_list_page.dart`: added an explicit bottom navigation bar (was implicit through `HomePage`)
* `product_page.dart`: added an explicit bottom navigation bar (was not there)
* `smooth_ui_library.dart`: removed "UI" version of smooth bottom navigation bar
* `user_preferences.dart`: added a preference for tab index
* `user_preferences_page.dart`: added an explicit bottom navigation bar (was implicit through `HomePage`)
