+++ packages/smooth_app/lib/main.dart
@@ -11,7 +11,7 @@ import 'package:smooth_app/data_models/product_preferences.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
 import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/database/search_history.dart';
-import 'package:smooth_app/pages/home_page.dart';
+import 'package:smooth_app/pages/smooth_bottom_navigation_bar.dart';
 import 'package:smooth_app/themes/smooth_theme.dart';
 import 'package:smooth_app/themes/theme_provider.dart';
 
@@ -171,7 +171,7 @@ class SmoothAppGetLanguage extends StatelessWidget {
       DefaultAssetBundle.of(context),
       languageCode,
     );
-    return const HomePage();
+    return SmoothBottomNavigationBar.getCurrentPage(context);
   }
 
   Future<void> _refresh(
