+++ packages/smooth_app/lib/pages/product/common/product_list_page.dart
@@ -10,6 +10,7 @@ import 'package:smooth_app/pages/personalized_ranking_page.dart';
 import 'package:smooth_app/pages/product/common/product_list_dialog_helper.dart';
 import 'package:smooth_app/pages/product/common/product_list_item.dart';
 import 'package:smooth_app/pages/product/common/product_query_page_helper.dart';
+import 'package:smooth_app/pages/smooth_bottom_navigation_bar.dart';
 import 'package:smooth_app/themes/smooth_theme.dart';
 
 class ProductListPage extends StatefulWidget {
@@ -87,6 +88,7 @@ class _ProductListPageState extends State<ProductListPage> {
         throw Exception('unknown list type ${productList.listType}');
     }
     return Scaffold(
+      bottomNavigationBar: SmoothBottomNavigationBar(),
       appBar: AppBar(
         backgroundColor: SmoothTheme.getColor(
           colorScheme,
