--- packages/smooth_ui_library/lib/navigation/models/smooth_bottom_navigation_bar_item.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SmoothBottomNavigationBarItem {
-  SmoothBottomNavigationBarItem(
-      {required this.name, required this.iconPath, required this.body});
-  String name;
-  String iconPath;
-  Widget body;
-}
